+++ src/ethereum/EthereumEngine.ts
@@ -527,6 +527,7 @@ export class EthereumEngine extends CurrencyEngine<
    */
   // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
   async updateNetworkFees() {
+    // Update network gasPrice:
     for (const externalFeeProvider of this.externalFeeProviders) {
       try {
         const ethereumFee = await externalFeeProvider()
@@ -550,6 +551,17 @@ export class EthereumEngine extends CurrencyEngine<
             externalFeeProvider.name
           }. ${JSON.stringify(e)}`
         )
+      }
+    }
+
+    // Update network baseFee:
+    if (this.networkInfo.supportsEIP1559 === true) {
+      try {
+        const baseFee = await this.ethNetwork.getBaseFeePerGas()
+        if (baseFee == null) return
+        this.networkFees.default.baseFee = baseFee
+      } catch (error: any) {
+        this.error(`Error fetching base fee: ${JSON.stringify(error)}`)
       }
     }
   }
@@ -694,10 +706,21 @@ export class EthereumEngine extends CurrencyEngine<
       this.networkInfo.feeUpdateFrequencyMs ?? NETWORK_FEES_POLL_MILLISECONDS
     // Fetch the static fees from the info server only once to avoid overwriting live values.
     this.infoFeeProvider()
-      .then(info => {
+      .then(async info => {
         this.log.warn(`infoFeeProvider:`, JSON.stringify(info, null, 2))
 
         this.networkFees = mergeDeeply(this.networkFees, info)
+
+        // Update network baseFee:
+        if (this.networkInfo.supportsEIP1559 === true) {
+          try {
+            const baseFee = await this.ethNetwork.getBaseFeePerGas()
+            if (baseFee == null) return
+            this.networkFees.default.baseFee = baseFee
+          } catch (error) {
+            this.error(`Error fetching base fee: ${JSON.stringify(error)}`)
+          }
+        }
       })
       .catch(() => this.warn('Error fetching fees from Info Server'))
       .finally(
@@ -918,7 +941,7 @@ export class EthereumEngine extends CurrencyEngine<
     const feeParams = await getFeeParamsByTransactionType(
       txType,
       toHex(miningFees.gasPrice),
-      this.ethNetwork.getBaseFeePerGas
+      this.networkFees.default.baseFee
     )
 
     //