+++ src/ethereum/info/ethereumInfo.ts
@@ -1004,6 +1004,7 @@ const builtinTokens: EdgeTokenMap = {
 
 const defaultNetworkFees: EthereumFees = {
   default: {
+    baseFee: undefined,
     baseFeeMultiplier: {
       lowFee: '1',
       standardFeeLow: '1.25',
@@ -1027,6 +1028,7 @@ const defaultNetworkFees: EthereumFees = {
     minPriorityFee: '2000000000'
   },
   '1983987abc9837fbabc0982347ad828': {
+    baseFee: undefined,
     baseFeeMultiplier: undefined,
     // @ts-expect-error
     gasLimit: {
@@ -1045,6 +1047,7 @@ const defaultNetworkFees: EthereumFees = {
     minPriorityFee: undefined
   },
   '2983987abc9837fbabc0982347ad828': {
+    baseFee: undefined,
     baseFeeMultiplier: undefined,
     // @ts-expect-error
     gasLimit: {